small change to prune function
I think this should be added because the prune function will also delete `tr!prune` command and if you say `tr!prune 5` it will delete 4 messages and the command but if we add this in then it will delete 5 messages and the command. It just makes more sense to me.